move degree plan and editor into student folder
