+++ client/src/app/student/s-degree-plan-editor/s-degree-plan-editor.component.spec.ts
