+++ client/src/app/student/s-degree-plan/s-degree-plan.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, ErrorHandler, Input } from "@angular/core";
 import { Router } from "@angular/router";
 import { Observable } from "rxjs";
-import { PlanService, Year } from "../plan.service";
+import { PlanService, Year } from "../../plan.service";
 
 @Component({
   selector: "app-s-degree-plan",
